refactor(a11y): deslop FP-fix rules (reuse helpers, de-nest, relocate getImplicitRole)
Behavior-preserving cleanup of the accessibility false-positive fixes:

- interactive-supports-focus: replace the inline spread-attribute check with
  the existing `hasJsxSpreadAttribute` util (drops a byte-for-byte
  reimplementation and a now-dead `isNodeOfType` import).
- no-redundant-roles: de-nest the implicit-role ternary into an if/else chain
  (AGENTS.md: no nested ternaries) and de-shadow the filter parameter.
- prefer-tag-over-role: merge the two disjoint role sets that were only ever
  used together in one boolean union into one `ROLES_WITHOUT_CLEAN_TAG`.
- role-supports-aria-props / no-redundant-roles: relocate `getImplicitRole`
  into `utils/get-implicit-role.ts`, removing the only rule-to-rule import in
  the a11y directory (shared helpers belong in utils/).
- control-has-associated-label: un-glue and tighten a comment block.

No behavior change: full oxlint-plugin suite (7380 tests) green; typecheck,
lint, and format all clean.
