+++ packages/oxlint-plugin-react-doctor/src/plugin/rules/a11y/prefer-tag-over-role.ts
@@ -8,34 +8,24 @@ import { getTagsForRole } from "../../constants/aria-element-roles.js";
 const buildMessage = (role: string, tag: string): string =>
   `Screen reader users get more reliable semantics from \`<${tag}>\` than \`role="${role}"\`, so use \`<${tag}>\` instead.`;
 
-// Composite-widget roles whose first "matching tag" is NOT a drop-in
-// standalone replacement: `role="listbox"` reverse-maps to `<datalist>`
-// / `<select>`, but `<datalist>` only works as an `<input list>`
-// autocomplete source and `<select>` is a native form control — neither
-// substitutes for a custom navigable results widget (wired via
-// aria-controls / aria-activedescendant). Same for `combobox`. `option`
-// is the child role of those widgets: a native `<option>` only renders
-// inside `<select>`/`<datalist>`/`<optgroup>` and accepts text only, so a
-// `<div role="option">` holding arbitrary JSX in a custom listbox cannot
-// become one. Suggesting the tag here is semantically wrong, so we don't
-// flag these roles.
-const COMPOSITE_WIDGET_ROLES: ReadonlySet<string> = new Set(["listbox", "combobox", "option"]);
-
-// Roles whose first reverse-mapped tag is wrong, semantically loaded, or
-// otherwise NOT a safe drop-in for a generic `div`/`span` container:
-//   - `group`   → `<address>` (contact-info element — nonsensical; the only
-//                 real `group` element is the form-specific `<fieldset>`).
-//   - `region`  → `<section>`, which only exposes the `region` role when it
-//                 carries an accessible name.
-//   - `img`     → `<img>`, a void element that cannot contain children; a
-//                 `<div|span role="img">` is the compose-an-image-from-non-img
-//                 -content idiom (icon font / inline SVG / spinner) with no src.
-//   - `dialog`  → `<dialog>`, which has top-layer/`.showModal()` behavior that
-//                 a portal+aria custom dialog can't adopt without a rewrite.
-//   - `status`  → `<output>`, a form-association element for calculation
-//                 results — not a drop-in for a generic live-region status.
-// Suggesting any of these would mislead, so skip the roles.
-const STRUCTURAL_ROLES_WITHOUT_CLEAN_TAG: ReadonlySet<string> = new Set([
+// Roles whose first reverse-mapped tag isn't a safe drop-in for a generic
+// `div`/`span`, so we decline to suggest it:
+//   - `listbox`/`combobox` → `<datalist>`/`<select>` (autocomplete source /
+//                 native form control, not a custom aria-controls widget).
+//   - `option`  → native `<option>` renders only inside `<select>`/`<datalist>`
+//                 and is text-only, so it can't hold arbitrary JSX.
+//   - `group`   → `<address>` (contact-info; the only real `group` element is
+//                 the form-specific `<fieldset>`).
+//   - `region`  → `<section>`, which exposes `region` only when named.
+//   - `img`     → `<img>` is void and can't wrap the compose-an-image content
+//                 (icon font / inline SVG / spinner) a `role="img"` div holds.
+//   - `dialog`  → `<dialog>` has top-layer/`.showModal()` behavior a portal+aria
+//                 custom dialog can't adopt without a rewrite.
+//   - `status`  → `<output>` is a form-result element, not a live-region status.
+const ROLES_WITHOUT_CLEAN_TAG: ReadonlySet<string> = new Set([
+  "listbox",
+  "combobox",
+  "option",
   "group",
   "region",
   "img",
@@ -82,7 +72,7 @@ export const preferTagOverRole = defineRule({
       if (!roleAttr) return;
       const role = getJsxPropStringValue(roleAttr);
       if (!role) return;
-      if (COMPOSITE_WIDGET_ROLES.has(role) || STRUCTURAL_ROLES_WITHOUT_CLEAN_TAG.has(role)) return;
+      if (ROLES_WITHOUT_CLEAN_TAG.has(role)) return;
       if (
         role === "separator" &&
         SPLITTER_SIGNAL_ATTRIBUTES.some((attribute) =>
